Fix OOM in WPF compositor by freezing RenderTargetBitmap and consolidating into XamlIconRenderer
RenderTargetBitmap holds open DUCE compositor channels until frozen or GC'd.
Without Freeze(), repeated icon renders (on DPI changes, context menu opens, etc.)
accumulated GPU resources until the compositor ran out, causing OutOfMemoryException
deep in DUCE.Channel.SyncFlush even when process heap looked fine.

Consolidates three identical private copies of RenderXaml (in WFListEditor,
WFEditorTree, WFAttributesControl) plus the WFToolbar call into a single
XamlIconRenderer utility class.
